fix(macos): align recommendedVMMemoryMiB return type with int64
strconv.ParseInt returns int64, so comparing against the int return
of recommendedVMMemoryMiB fails to compile on darwin. Widen the
helper to int64 and swap strconv.Itoa for strconv.FormatInt at the
single caller that formats it for the podman machine set --memory
argument. Linux builds never saw the error because startstop_darwin.go
is darwin-tagged.

+++ internal/cli/podman_memory.go
@@ -3,7 +3,8 @@ package cli
 // recommendedVMMemoryMiB picks Podman Machine memory based on host RAM.
 // Returns 4096 (safe default) when detection fails so we never go below
 // the floor that fixes the v1.12.6 MySQL+PHP-FPM+Horizon swap thrash.
-func recommendedVMMemoryMiB(hostMemoryGiB int) int {
+// int64 matches strconv.ParseInt at the comparison site.
+func recommendedVMMemoryMiB(hostMemoryGiB int) int64 {
 	switch {
 	case hostMemoryGiB <= 0:
 		return 4096
